🔀Fix float to int

+++ solathon/utils.py
@@ -1,6 +1,6 @@
-def lamport_to_sol(lamports: float) -> float:
-    return float(lamports / 1000000000)
+def lamport_to_sol(lamports: int) -> int:
+    return int(lamports / 1000000000)
 
-def sol_to_lamport(lamports: float) -> float:
-    return float(lamports * 1000000000)
+def sol_to_lamport(lamports: int) -> int:
+    return int(lamports * 1000000000)
